+++ NEWS.md
@@ -1,3 +1,10 @@
+# effectsize 0.4.0.001
+
+## New features
+
+- `standardize()` gains support for `mediation::mediate()` models.
+- `eta_squared()` family available for `manova` objects.
+
 # effectsize 0.4.0
 
 ## Breaking Changes